refactor(engine): discussion-map transactions move into the domain ring
The two discussion-map verb bodies assembled lock→load→apply→save inline in
engine.cjs, unlike every other noun whose domain module owns its lock. New
recordSubtopicAdd/recordSubtopicState transactions in domain/discussion-map.cjs;
the CLI dispatch is now a thin call and engine.cjs no longer imports the
kernel manifest facade. Byte-identical CLI behaviour.
